backend/labs/4/scripts/analyzy/requirements.txt
asttokens==3.0.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
fonttools==4.60.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.5
threadpoolctl==3.6.0
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.14
